requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.36
psycopg[binary]>=3.3.2
alembic>=1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Google Gemini AI
google-generativeai==0.3.2

# Validation and serialization
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
